--- .eslintrc.json
@@ -1,26 +1,0 @@
-{
-    "env": {
-      "browser": true,
-      "node": true,
-      "commonjs": true,
-      "jest": true,
-      "es6": true
-    },
-    "globals": {
-      "fail":true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-      "sourceType": "module",
-      "ecmaVersion": 2018
-    },
-    "rules": {
-      "no-console": "off",
-      "indent": [ "error", 2 ],
-      "quotes": ["error", "single", { "allowTemplateLiterals": true }],
-      "comma-dangle": ["error", "always-multiline"],
-      "semi": [ "error", "always" ],
-      "no-unused-vars": [1, {"vars": "local", "args": "none"}]
-    }
-  }
-
