+++ .github/workflows/npm-publish.yml
@@ -1,6 +1,6 @@
-name: Node CI
+name: Run Javascript Tests
 
-on: [push,pull_request]
+on: [push]
 
 jobs:
   build:
@@ -9,18 +9,17 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [10.x]
+        node-version: [12.x]
 
     steps:
     - uses: actions/checkout@v1
-    - name: Use Node.js $
+    - name: Use Node.js ${{ matrix.node-version }}
       uses: actions/setup-node@v1
       with:
-        node-version: $
+        node-version: ${{ matrix.node-version }}
     - name: npm install, build, and test
       run: |
-        npm ci
-        npm run build --if-present
+        npm i
         npm test
       env:
         CI: true
