+++ src/auth/middleware/basic.js
@@ -12,10 +12,10 @@ const Users = require('../models/users-model.js');
 
 module.exports = async(request, response, next) => {
 
-    let basicHeaderParts = req.headers.authorization.split(' '); // ['Basic', 'sdkjdsljd=']
-    let encodedString = basicHeaderParts.pop(); // sdkjdsljd=
-    let decodedString = base64.decode(encodedString); // "username:password"
-    let [username, password] = decodedString.split(':'); // username, password
+    let basicHeaderParts = req.headers.authorization.split(' ');
+    let encodedString = basicHeaderParts.pop();
+    let decodedString = base64.decode(encodedString);
+    let [username, password] = decodedString.split(':');
 
     try {
         const user = await Users.findOne({ username: username })
